Systemd unit: Use bluetooth.service instead of bluetooth.target
In my understanding, targets are meant for saying “start this unit when the target is requested”, i.e. for the “WantedBy” setting.

Otherwise, dependencies are implemented using services. That means we should be using the “bluetooth.service”, not the bluetooth target.

Note that in my environment (Ubuntu 20.04, 64bit) this worked with both configs, so it’d be ideal if you can test in your Raspbian environment.